pyproject.toml
[tool.poetry]
name = "hugit"
version = "0.0.1"
description = "Hugit"
authors = ["Daniel van Strien <daniel.van-strien@bl.uk>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/davanstrien/hugit"
repository = "https://github.com/davanstrien/hugit"
documentation = "https://hugit.readthedocs.io"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/davanstrien/hugit/releases"

[tool.poetry.dependencies]
python = ">=3.7 <4.0.0"
toolz = "^0.11.2"
rich = "^12.0.1"
Pillow = "^9.0.1"
attrs = "^21.4.0"
typing-extensions = "^4.1.1"
typed-settings = "^1.0.0"
datasets = "^2.0.0"

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = "^0.17.0"

[tool.poetry.scripts]
hugit = "hugit.cli:cli"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["hugit" ]

[tool.coverage.report]
show_missing = true
fail_under = 90

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = false
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true

[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
